refactors ghost_examples to ghost_example_pkg
