+++ 10_Examples/ghost_example_pkg/config/example.yaml
