+++ 10_Examples/ghost_example_pkg/launch/pub_sub_example.launch.py
@@ -6,14 +6,14 @@ def generate_launch_description():
 
     publisher_node = launch_ros.actions.Node(
         name="publisher_node",
-        package="ghost_examples",
+        package="ghost_example_pkg",
         # This name is specified in the CmakeLists.txt file of this package
         executable="ros_publisher_example",
     )
 
     subscriber_node = launch_ros.actions.Node(
         name="subscriber_node",
-        package="ghost_examples",
+        package="ghost_example_pkg",
         # This name is specified in the CmakeLists.txt file of this package
         executable="ros_subscriber_example_main",
     )
