+++ 10_Examples/ghost_example_pkg/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>ghost_examples</name>
+  <name>ghost_example_pkg</name>
   <version>0.0.0</version>
   <description>Provides example code for new VEXU team members</description>
   <maintainer email="jessemaxxwilson@utexas.edu">maxx</maintainer>
