+++ 10_Examples/ghost_example_pkg/scripts/cubic_trajectory.py
