+++ 10_Examples/ghost_example_pkg/src/ros_publisher_example.cpp
