+++ 10_Examples/ghost_example_pkg/urdf/cylinder.xacro
